Policies and Gates

+++ app/Http/Controllers/TareaController.php
@@ -152,7 +152,10 @@ class TareaController extends Controller
     public function destroy(Tarea $tarea)
     {
         //
-        $tarea->delete();
-        return redirect()->route('tarea.index');
+        if(\Gate::allows('administrador',$tarea)){
+            $tarea->delete();
+            return redirect()->route('tarea.index');
+        }
+        
     }
 }
